Move all internal code to `vercel._internal`
