+++ src/vercel/_internal/blob/__init__.py
@@ -7,11 +7,9 @@ import uuid
 from collections.abc import Awaitable, Callable, Iterable
 from dataclasses import dataclass
 from datetime import datetime
-from typing import Any, Literal, Protocol, TypedDict
+from typing import Any, Protocol, TypedDict
 
-from .errors import BlobError, BlobNoTokenProvidedError
-
-Access = Literal["public", "private"]
+from vercel.blob.errors import BlobError, BlobNoTokenProvidedError
 
 DEFAULT_VERCEL_BLOB_API_URL = "https://vercel.com/api/blob"
 MAXIMUM_PATHNAME_LENGTH = 950
@@ -109,13 +107,13 @@ def validate_path(path: str) -> None:
             raise BlobError(f'path cannot contain "{invalid}", please encode it if needed')
 
 
-def validate_access(access: Access) -> Access:
+def validate_access(access: str) -> str:
     if access not in ("public", "private"):
         raise BlobError('access must be "public" or "private"')
     return access
 
 
-def construct_blob_url(store_id: str, pathname: str, access: Access) -> str:
+def construct_blob_url(store_id: str, pathname: str, access: str) -> str:
     """Construct a blob storage URL based on access type.
 
     Public:  https://{storeId}.public.blob.vercel-storage.com/{pathname}
@@ -150,19 +148,6 @@ def compute_body_length(body: Any) -> int:
     return 0
 
 
-# Progress
-@dataclass
-class UploadProgressEvent:
-    loaded: int
-    total: int
-    percentage: float
-
-
-OnUploadProgressCallback = (
-    Callable[[UploadProgressEvent], None] | Callable[[UploadProgressEvent], Awaitable[None]]
-)
-
-
 class SupportsRead(Protocol):
     def read(self, size: int = -1) -> bytes:  # pragma: no cover - Protocol
         ...
@@ -178,7 +163,7 @@ class StreamingBodyWithProgress:
     def __init__(
         self,
         body: bytes | bytearray | memoryview | str | SupportsRead | Iterable[bytes],
-        on_progress: OnUploadProgressCallback | None,
+        on_progress: Callable | None,
         chunk_size: int = 64 * 1024,
         total: int | None = None,
     ) -> None:
@@ -229,6 +214,8 @@ class StreamingBodyWithProgress:
 
     def _emit_progress(self) -> None:
         if self._on_progress:
+            from vercel.blob.types import UploadProgressEvent
+
             total = self._total if self._total else self._loaded
             percentage = round((self._loaded / total) * 100, 2) if total else 0.0
             self._on_progress(
@@ -237,6 +224,8 @@ class StreamingBodyWithProgress:
 
     async def _emit_progress_async(self) -> None:
         if self._on_progress:
+            from vercel.blob.types import UploadProgressEvent
+
             total = self._total if self._total else self._loaded
             percentage = round((self._loaded / total) * 100, 2) if total else 0.0
             result = self._on_progress(
@@ -313,30 +302,6 @@ def parse_datetime(value: str) -> datetime:
         return datetime.strptime(value, "%Y-%m-%dT%H:%M:%S.%fZ")
 
 
-def get_download_url(blob_url: str) -> str:
-    try:
-        from urllib.parse import parse_qsl, urlencode, urlparse, urlunparse
-
-        parsed = urlparse(blob_url)
-        q = dict(parse_qsl(parsed.query))
-        q["download"] = "1"
-        new_query = urlencode(q)
-        return urlunparse(
-            (
-                parsed.scheme,
-                parsed.netloc,
-                parsed.path,
-                parsed.params,
-                new_query,
-                parsed.fragment,
-            )
-        )
-    except Exception:
-        # Fallback: naive append
-        sep = "&" if "?" in blob_url else "?"
-        return f"{blob_url}{sep}download=1"
-
-
 # TypedDict with real HTTP header keys. Use functional syntax to allow hyphens.
 PutHeaders = TypedDict(
     "PutHeaders",
@@ -356,7 +321,7 @@ def create_put_headers(
     add_random_suffix: bool | None = None,
     allow_overwrite: bool | None = None,
     cache_control_max_age: int | None = None,
-    access: Access | None = None,
+    access: str | None = None,
 ) -> PutHeaders:
     headers: PutHeaders = {}
     if content_type:
