+++ src/vercel/_internal/telemetry/__init__.py
@@ -1,0 +1,6 @@
+"""Telemetry functionality for Vercel Python SDK (internal use)."""
+
+from vercel._internal.telemetry.client import TelemetryClient
+from vercel._internal.telemetry.tracker import track
+
+__all__ = ["TelemetryClient", "track"]
