--- src/vercel/_telemetry/__init__.py
@@ -1,6 +1,0 @@
-"""Telemetry functionality for Vercel Python SDK (internal use)."""
-
-from .client import TelemetryClient
-from .tracker import track
-
-__all__ = ["TelemetryClient", "track"]
