+++ src/vercel/blob/aio.py
@@ -42,7 +42,7 @@ from .types import (
     ListBlobResult,
     PutBlobResult,
 )
-from .utils import Access, OnUploadProgressCallback, UploadProgressEvent, get_download_url
+from .types import Access, OnUploadProgressCallback, UploadProgressEvent, get_download_url
 
 __all__ = [
     # errors
