+++ src/vercel/blob/multipart/api.py
@@ -4,19 +4,17 @@ import inspect
 from collections.abc import Awaitable, Callable
 from typing import Any, cast
 
-from ..._iter_coroutine import iter_coroutine
-from ..errors import BlobError
-from ..types import MultipartCreateResult, MultipartPart, PutBlobResult
-from ..utils import (
-    Access,
+from vercel._internal.iter_coroutine import iter_coroutine
+from vercel.blob.errors import BlobError
+from vercel.blob.types import Access, MultipartCreateResult, MultipartPart, PutBlobResult, UploadProgressEvent
+from vercel._internal.blob import (
     PutHeaders,
-    UploadProgressEvent,
     create_put_headers,
     ensure_token,
     validate_access,
     validate_path,
 )
-from .core import _AsyncMultipartClient, _BaseMultipartClient, _SyncMultipartClient
+from vercel._internal.blob.multipart import _AsyncMultipartClient, _BaseMultipartClient, _SyncMultipartClient
 
 
 def _validate_multipart_context(path: str, access: Access, token: str | None) -> str:
