+++ src/vercel/cache/cache_build.py
@@ -5,7 +5,7 @@ from collections.abc import Callable, Mapping, Sequence
 
 import httpx
 
-from .._telemetry.tracker import track
+from vercel._internal.telemetry.tracker import track
 from .types import AsyncCache, Cache
 
 HEADERS_VERCEL_CACHE_STATE = "x-vercel-cache-state"
