+++ tests/integration/test_http_transport_raw_body.py
@@ -4,14 +4,14 @@ import pytest
 import respx
 from httpx import Response
 
-from vercel._http import (
+from vercel._internal.http import (
     AsyncTransport,
     RawBody,
     SyncTransport,
     create_base_async_client,
     create_base_client,
 )
-from vercel._iter_coroutine import iter_coroutine
+from vercel._internal.iter_coroutine import iter_coroutine
 
 
 class TestRawBodySupport:
